+++ src/main/java/de/telran/onlineshopgarden/repository/CartItemRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 
     @Modifying
-    void deleteByProductId(Integer productId);
+    void deleteByProductId(Long productId);
 }
