+++ src/main/java/de/telran/onlineshopgarden/repository/OrderRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     List<Order> findByStatusNotIn(List<OrderStatus> statuses);
 }
